Move shared code together
